pr2/requirement.txt
certifi==2022.12.7
charset-normalizer==3.1.0
docutils==0.19
idna==3.4
kaki==0.1.8
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
Pillow==9.4.0
Pygments==2.14.0
pypiwin32==223
python-firebase==1.2
pywin32==306
requests==2.28.2
urllib3==1.26.15
watchdog==3.0.0
